backend/requirements.txt
# --- Flask core ---
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Flask-Session==0.5.0
Flask-Admin==1.6.1
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.6.0

# --- Database / ORM ---
SQLAlchemy==2.0.31
# psycopg2-binary removed for local Windows dev (use SQLite). Add back for production on Linux.

# --- Caching / Queue / Redis ---
redis==5.0.3
Celery==5.3.6

# --- Environment / Config ---
python-dotenv==1.0.1
itsdangerous==2.2.0

# --- File uploads / Cloud ---
cloudinary==1.40.0
Pillow==10.4.0

# --- Payment Gateways ---
stripe==9.12.0
paypalrestsdk==1.13.1
requests==2.32.3

# --- Utilities & Validation ---
email-validator==2.2.0

# --- Security / Encryption ---
cryptography==43.0.0

# --- Logging / Monitoring ---
blinker==1.8.2

# --- Production Server ---
gunicorn==23.0.0

# --- Optional Tools (safe to keep) ---
Flask-Cors==4.0.0
Flask-Limiter==3.6.0
